requirements.txt
certifi
charset-normalizer
contourpy
cycler
filelock
fonttools
fsspec
idna
jinja2
kiwisolver
lap
markupsafe
matplotlib
mpmath
networkx
numpy
opencv-python
opencv-python-headless
packaging
pandas
pillow
pip
polars
polars-runtime-32
psutil
pyparsing
python-dateutil
pyyaml
requests
scipy
setuptools
six
sympy
torch
torchaudio
torchvision
typing-extensions
tzdata
ultralytics
ultralytics-thop
urllib3
